feat: add common strings to i18n service
