+++ index.js
@@ -6,7 +6,7 @@ import { enableScreens } from 'react-native-screens'
 import App from './src/App'
 import * as i18n from './src/services/i18n'
 
-i18n.initialise()
+i18n.initialize()
 enableScreens()
 
 AppRegistry.registerComponent( 'App', () => App )
