+++ src/services/i18n/common-strings.ts
@@ -1,0 +1,9 @@
+import registerTranslations from './register-translations'
+
+const commonStrings = registerTranslations( {
+  done: {
+    'en-US': 'Done',
+  },
+} )
+
+export default commonStrings
